+++ src/datasets.js
@@ -191,6 +191,37 @@ export default {
     },
 
     /**
+     * Updates dataset.
+     *
+     * @memberof ApifyClient.datasets
+     * @instance
+     * @param {Object} options
+     * @param options.token
+     * @param {String} options.datasetId - Unique dataset ID
+     * @param {Object} options.dataset
+     * @param callback
+     * @returns {Dataset}
+     */
+    updateDataset: (requestPromise, options) => {
+        const { baseUrl, token, datasetId, dataset } = options;
+
+        checkParamOrThrow(baseUrl, 'baseUrl', 'String');
+        checkParamOrThrow(token, 'token', 'String');
+        checkParamOrThrow(datasetId, 'datasetId', 'String');
+        checkParamOrThrow(dataset, 'dataset', 'Object');
+
+        return requestPromise({
+            url: `${baseUrl}${BASE_PATH}/${datasetId}`,
+            json: true,
+            method: 'PUT',
+            qs: { token },
+            body: _.omit(dataset, 'id'),
+        })
+            .then(pluckData)
+            .then(parseDateFields);
+    },
+
+    /**
      * Deletes given dataset.
      *
      * @memberof ApifyClient.datasets
@@ -236,6 +267,8 @@ export default {
      *   Otherwise they are sorted in ascending order.
      * @param {Array} [options.fields]
      *   An array of field names that will be included in the result. If omitted, all fields are included in the results.
+     * @param {Array} [options.omit]
+     *   An array of field names that will be removed from the result. If omitted, all fields are included in the results.
      * @param {String} [options.unwind]
      *   Specifies a name of the field in the result objects that will be used to unwind the resulting objects.
      *   By default, the results are returned as they are.
@@ -295,7 +328,7 @@ export default {
         checkParamOrThrow(options.offset, 'offset', 'Maybe Number');
         checkParamOrThrow(options.limit, 'limit', 'Maybe Number');
         checkParamOrThrow(options.fields, 'fields', 'Maybe [String]');
-        checkParamOrThrow(options.omit, 'omit', 'Maybe Array');
+        checkParamOrThrow(options.omit, 'omit', 'Maybe [String]');
         checkParamOrThrow(options.delimiter, 'delimiter', 'Maybe String');
         checkParamOrThrow(options.unwind, 'unwind', 'Maybe String');
         checkParamOrThrow(options.xmlRoot, 'xmlRoot', 'Maybe String');
@@ -314,7 +347,7 @@ export default {
         checkParamOrThrow(options.skipFailedPages, 'skipFailedPages', 'Maybe Boolean');
 
         // Pick query params.
-        const query = _.pick(options, 'offset', 'limit', 'fields', 'omit', 'delimiter', 'unwind', 'xmlRoot', 'xmlRow', 'format', 'token');
+        const query = _.pick(options, 'offset', 'limit', 'delimiter', 'unwind', 'xmlRoot', 'xmlRow', 'format', 'token');
 
         // Add Boolean query params.
         if (options.skipHeaderRow) query.skipHeaderRow = 1;
@@ -330,7 +363,8 @@ export default {
         if (options.bom) query.bom = 1;
         else if (options.bom === false) query.bom = 0;
 
-        if (query.fields) query.fields = query.fields.join(',');
+        if (options.fields) query.fields = options.fields.join(',');
+        if (options.omit) query.omit = options.omit.join(',');
         const requestOpts = {
             url: `${baseUrl}${BASE_PATH}/${datasetId}/items`,
             method: 'GET',